\name{MAGNAMWAR-package}
\alias{MAGNAMWAR-package}
\alias{MAGNAMWAR}
\docType{package}
\title{
\packageTitle{MAGNAMWAR}
}
\description{
\packageDescription{MAGNAMWAR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MAGNAMWAR}
\packageIndices{MAGNAMWAR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MAGNAMWAR}

Maintainer: \packageMaintainer{MAGNAMWAR}
}
\references{
~~ Literature or other references for background information ~~
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
