% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{pheno_data}
\alias{pheno_data}
\title{Development of fruit flies dataset.}
\format{A data frame with 586 rows and 4 variables:
\describe{
  \item{Treatment}{4-letter taxa designation of associated bacteria}
  \item{RespVar}{response variable, }
  \item{Vial}{random effect variable, vial number of flies}
  \item{Experiment}{random effect variable, experiment number of flies}
}}
\usage{
pheno_data
}
\description{
A subset of the development of fruit flies, collected in the Chaston Lab,
to be used as a brief example for survival tests in analyzeOrthoMCL.
}
\keyword{datasets}

