% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_repseq.R
\name{pick_repseq}
\alias{pick_repseq}
\title{Pick Representative Sequences}
\usage{
pick_repseq(mcl_data, fa_dir)
}
\arguments{
\item{mcl_data}{output of format_afterOrtho() --list of 2 things-- 1: binary matrix indicating the presence / absence of genes in each COG and 2: vector of names of COGs}

\item{fa_dir}{Path to the directory where all raw GenBank files are stored. Note, all file names must be changed to a 4-letter code representing each species and have '.fasta' file descriptor}
}
\value{
The path to the representative sequence file in fasta format
}
\description{
Randomly picks a representative sequence from GenBank fasta files for every OrthoMCL COG
}
\examples{
dir <- system.file('sample_data', 'fasta_dir', package='MAGNAMWAR')
dir <- paste(dir,'/',sep='')
repseqs <- pick_repseq(after_ortho_format, dir)
}

