% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printCOGseqs.R
\name{printCOGseqs}
\alias{printCOGseqs}
\title{Print COG Sequences}
\usage{
printCOGseqs(after_ortho, COG, fasta_dir, out_dir = NULL, outfile = "none")
}
\arguments{
\item{after_ortho}{output from format_afterOrtho}

\item{COG}{name of COG}

\item{fasta_dir}{directory to fastas}

\item{out_dir}{complete path to output directory}
}
\value{
A fasta file with all protein sequnces and ids for a given COG
}
\description{
Print all protein sequences and annotations in a given COG
}
\examples{

COG <- 'OG5_126968'
dir <- system.file('sample_data', 'fasta_dir', package='MAGNAMWAR')
dir <- paste(dir,'/',sep='')
 
printCOGseqs(after_ortho_format,COG,dir)

}

