% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_append_matrix.R
\name{surv_append_matrix}
\alias{surv_append_matrix}
\title{Append Survival Test Outputs}
\usage{
surv_append_matrix(work_dir, out_name = "surv_matrix.csv", out_dir = NULL)
}
\arguments{
\item{work_dir}{the directory where the output files of analyze_OrthoMCL are located}

\item{out_name}{file name of outputted matrix}

\item{out_dir}{the directory where the outputted matrix is placed}
}
\value{
A csv file containing a matrix with the following columns: COG, p-values, Bonferroni corrected p-values, mean phenotype of COG-containing taxa, mean pheotype of COG-lacking taxa, taxa included in COG, taxa not included in COG
}
\description{
Function used to append all .csv files that are outputted from analyze_OrthoMCL into one matrix.
}
\examples{
file <- system.file('sample_data', 'outputs', package='MAGNAMWAR')
directory <- paste(file, '/', sep = '')
surv_append_matrix(directory)


}
\references{
Some sort of reference
}

